refactor: split fifth app.rs submodules under ADR-028
